Move tensor to array functions to utils (#7810)
* Move tensor to array functions to utils

* fix lint errors
