+++ tfjs-layers/src/layers/nlp/tokenizers_utils_test.ts
@@ -19,8 +19,7 @@ import { tensor, test_util } from '@tensorflow/tfjs-core';
 
 import { BytePairTokenizerCache, SPLIT_PATTERN_1, bytesToUnicode,
   createAltsForUnsplittableTokens, createStaticHashtable, regexSplit,
-  removeStringsFromInputs, splitStringsForBpe, tensorArrTo2DArr,
-  tensorToArr } from './tokenizers_utils';
+  removeStringsFromInputs, splitStringsForBpe } from './tokenizers_utils';
 import { expectTensorsClose } from '../../utils/test_utils';
 
 describe('bytesToUnicode', () => {
@@ -229,26 +228,5 @@ describe('splitStringsForBpe', () => {
     expect(result.length).toBe(2);
     expectTensorsClose(result[0], tensor(['brown', '.']));
     expectTensorsClose(result[1], tensor(['black', '.']));
-  });
-});
-
-describe('tensor to array functions', () => {
-  it('tensorToArr', () => {
-    const inputStr = tensor(['these', 'are', 'strings', '.']);
-    const inputNum = tensor([2, 11, 15]);
-
-    test_util.expectArraysEqual(
-      tensorToArr(inputStr) as string[], ['these', 'are', 'strings', '.']);
-    test_util.expectArraysEqual(tensorToArr(inputNum) as number[], [2, 11, 15]);
-  });
-
-  it('tensorArrTo2DArr', () => {
-    const inputStr = [tensor(['these', 'are']), tensor(['strings', '.'])];
-    const inputNum = [tensor([2, 11]), tensor([15])];
-
-    test_util.expectArraysEqual(
-      tensorArrTo2DArr(inputStr) as string[][], [['these', 'are'], ['strings', '.']]);
-    test_util.expectArraysEqual(
-      tensorArrTo2DArr(inputNum) as number[][], [[2, 11], [15]]);
   });
 });
